requirements.txt
nicegui>=1.4.0
pystray>=0.19.0
python-socketio[client]>=5.11.0
python-escpos>=3.0a9
requests>=2.31.0
Pillow>=10.0.0
pyinstaller>=6.0.0
python-dotenv>=1.0.0
